Refaktoroi: Siirrä logiikkaa mockeissa uudelleenkäytettäväksi;
Tämä päättely ohitetaan nykyisellään käyttöoikeus-Mockeissa, mikä on johtanut esim.
siihen, että meillä on Mock-käyttäjinä käyttöoikeussettejä, joita ei pysty oikeassa
järjestelmässä olemaan.
