+++ src/main/scala/fi/oph/koski/userdirectory/DirectoryClient.scala
@@ -2,8 +2,10 @@ package fi.oph.koski.userdirectory
 
 import com.typesafe.config.Config
 import fi.oph.koski.cache.{CacheManager, Cached, CachingProxy, ExpiringCache}
-import fi.oph.koski.koskiuser.Käyttöoikeus
+import fi.oph.koski.koskiuser.{Käyttöoikeus, KäyttöoikeusGlobal, KäyttöoikeusOrg, KäyttöoikeusViranomainen, Palvelurooli, Rooli}
 import fi.oph.koski.log.NotLoggable
+import fi.oph.koski.organisaatio.Opetushallitus
+import fi.oph.koski.schema.OidOrganisaatio
 import fi.oph.koski.sso.CasService
 
 import scala.concurrent.duration.DurationInt
@@ -23,6 +25,26 @@ object DirectoryClient {
       case _ => new OpintopolkuDirectoryClient(config, casService)
     })
   }
+
+  def resolveKäyttöoikeudet(käyttäjä: HenkilönKäyttöoikeudet): (String, List[Käyttöoikeus]) =
+    (käyttäjä.oidHenkilo, käyttäjä.organisaatiot.flatMap {
+      case OrganisaatioJaKäyttöoikeudet(organisaatioOid, käyttöoikeudet) =>
+        val roolit = käyttöoikeudet.collect { case PalveluJaOikeus(palvelu, oikeus) => Palvelurooli(palvelu, oikeus) }
+        if (!roolit.map(_.palveluName).exists(List("KOSKI", "VALPAS").contains)) {
+          Nil
+        } else if (organisaatioOid == Opetushallitus.organisaatioOid) {
+          List(KäyttöoikeusGlobal(roolit))
+        } else if (hasViranomaisRooli(roolit)) {
+          List(KäyttöoikeusViranomainen(roolit))
+        } else {
+          List(KäyttöoikeusOrg(OidOrganisaatio(organisaatioOid), roolit, juuri = true, oppilaitostyyppi = None))
+        }
+    })
+
+  private def hasViranomaisRooli(roolit: List[Palvelurooli]) =
+    roolit.exists(r => Rooli.globaalitKoulutusmuotoRoolit.contains(r.rooli)) ||
+      roolit.map(_.rooli).contains(Rooli.TIEDONSIIRTO_LUOVUTUSPALVELU)
+
 }
 
 case class DirectoryUser(oid: String, käyttöoikeudet: List[Käyttöoikeus], etunimet: String, sukunimi: String, asiointikieli: Option[String])
